👔 Add DeliveryInfo, OrderGroup And Update
테스트 코드 작성 및 Dto들 수정

+++ src/main/java/com/example/orderapi/service/deliveryinfo/DeliveryInfoServiceImpl.java
@@ -1,6 +1,9 @@
 package com.example.orderapi.service.deliveryinfo;
 
-import com.example.orderapi.dto.deliveryinfo.*;
+import com.example.orderapi.dto.deliveryinfo.DeliveryInfoArrivedAtUpdateRequest;
+import com.example.orderapi.dto.deliveryinfo.DeliveryInfoCreateRequest;
+import com.example.orderapi.dto.deliveryinfo.DeliveryInfoLogisticsUpdateRequest;
+import com.example.orderapi.dto.deliveryinfo.DeliveryInfoResponse;
 import com.example.orderapi.entity.deliveryinfo.DeliveryInfo;
 import com.example.orderapi.repository.deliveryinfo.DeliveryInfoRepository;
 import jakarta.transaction.Transactional;
@@ -10,14 +13,13 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 @RequiredArgsConstructor
-@Service
 @Transactional
+@Service
 public class DeliveryInfoServiceImpl implements DeliveryInfoService {
     private final DeliveryInfoRepository deliveryInfoRepository;
 
-
     @Override
-    public DeliveryInfoResponse getById(Long id) {
+    public DeliveryInfoResponse getDeliveryInfoById(Long id) {
         Optional<DeliveryInfo> optionalDeliveryInfo = deliveryInfoRepository.findById(id);
         if (optionalDeliveryInfo.isEmpty()) {
             throw new RuntimeException();
@@ -27,54 +29,40 @@ public class DeliveryInfoServiceImpl implements DeliveryInfoService {
     }
 
     @Override
-    public DeliveryInfoResponse create(DeliveryInfoCreateRequest deliveryInfoCreateRequest) {
+    public DeliveryInfoResponse createDeliveryInfo(DeliveryInfoCreateRequest deliveryInfoCreateRequest) {
         DeliveryInfo deliveryInfo = new DeliveryInfo();
-        deliveryInfo.setName(deliveryInfoCreateRequest.getName());
-        deliveryInfo.setInvoiceNumber(deliveryInfoCreateRequest.getInvoiceNumber());
-        deliveryInfoRepository.save(deliveryInfo);
-        return DeliveryInfoResponse.fromEntity(deliveryInfoRepository.save(deliveryInfo));
+        deliveryInfo.ofCreate(deliveryInfoCreateRequest.getName(), deliveryInfoCreateRequest.getInvoiceNumber());
+        DeliveryInfo savedDeliveryInfo = deliveryInfoRepository.save(deliveryInfo);
+        return DeliveryInfoResponse.fromEntity(savedDeliveryInfo);
     }
 
     @Override
-    public DeliveryInfoResponse updateAll(Long id, DeliveryInfoUpdateRequest deliveryInfoUpdateRequest) {
+    public DeliveryInfoResponse updateDeliveryInfoLogistics(Long id, DeliveryInfoLogisticsUpdateRequest deliveryInfoLogisticsUpdateRequest) {
         Optional<DeliveryInfo> optionalDeliveryInfo = deliveryInfoRepository.findById(id);
         if (optionalDeliveryInfo.isEmpty()) {
             throw new RuntimeException();
         }
         DeliveryInfo deliveryInfo = optionalDeliveryInfo.get();
-        deliveryInfo.setDeliveryDate(deliveryInfoUpdateRequest.getDeliveryDate());
-        deliveryInfo.setForwardedAt(deliveryInfoUpdateRequest.getForwardedAt());
-        deliveryInfo.setArrivedAt(deliveryInfoUpdateRequest.getArrivedAt());
-        DeliveryInfo updateDeliverInfo = deliveryInfoRepository.save(deliveryInfo);
-        return DeliveryInfoResponse.fromEntity(updateDeliverInfo);
-    }
-
-    @Override
-    public DeliveryInfoResponse updateLogistics(Long id, DeliveryInfoLogisticsUpdateRequest deliveryInfoLogisticsUpdateRequest) {
-        Optional<DeliveryInfo> optionalDeliveryInfo = deliveryInfoRepository.findById(id);
-        if (optionalDeliveryInfo.isEmpty()) {
-            throw new RuntimeException();
-        }
-        DeliveryInfo deliveryInfo = optionalDeliveryInfo.get();
-        deliveryInfo.setForwardedAt(deliveryInfoLogisticsUpdateRequest.getForwardedAt());
-        deliveryInfo.setDeliveryDate(deliveryInfoLogisticsUpdateRequest.getDeliveryDate());
-        DeliveryInfo updateDeliverInfo = deliveryInfoRepository.save(deliveryInfo);
-        return DeliveryInfoResponse.fromEntity(updateDeliverInfo);
-    }
-
-    @Override
-    public DeliveryInfoResponse updateArrivedAt(Long id, DeliveryInfoArrivedAtUpdateRequest deliveryInfoArrivedAtUpdateRequest) {
-        Optional<DeliveryInfo> optionalDeliveryInfo = deliveryInfoRepository.findById(id);
-        if (optionalDeliveryInfo.isEmpty()) {
-            throw new RuntimeException();
-        }
-        DeliveryInfo deliveryInfo = optionalDeliveryInfo.get();
-        deliveryInfo.setArrivedAt(deliveryInfoArrivedAtUpdateRequest.getArrivedAt());
+        deliveryInfo.ofUpdateLogistics(deliveryInfoLogisticsUpdateRequest.getForwardedAt(), deliveryInfoLogisticsUpdateRequest.getDeliveryDate());
         return DeliveryInfoResponse.fromEntity(deliveryInfo);
     }
 
     @Override
-    public void delete(Long id) {
+    public DeliveryInfoResponse updateDeliveryInfoArrivedAt(Long id, DeliveryInfoArrivedAtUpdateRequest deliveryInfoArrivedAtUpdateRequest) {
+        Optional<DeliveryInfo> optionalDeliveryInfo = deliveryInfoRepository.findById(id);
+        if (optionalDeliveryInfo.isEmpty()) {
+            throw new RuntimeException();
+        }
+        DeliveryInfo deliveryInfo = optionalDeliveryInfo.get();
+        deliveryInfo.ofUpdateArrived(deliveryInfoArrivedAtUpdateRequest.getArrivedAt());
+        return DeliveryInfoResponse.fromEntity(deliveryInfo);
+    }
+
+    @Override
+    public void deleteDeliveryInfo(Long id) {
+        if (!deliveryInfoRepository.existsById(id)) {
+            throw new RuntimeException();
+        }
         deliveryInfoRepository.deleteById(id);
     }
 }
