+++ src/main/java/com/example/orderapi/service/ordergroup/OrderGroupServiceImpl.java
@@ -1,10 +1,16 @@
 package com.example.orderapi.service.ordergroup;
 
+import com.example.orderapi.dto.deliveryinfo.DeliveryInfoResponse;
 import com.example.orderapi.dto.ordergroup.OrderGroupCreateRequest;
 import com.example.orderapi.dto.ordergroup.OrderGroupResponse;
-import com.example.orderapi.dto.ordergroup.OrderGroupUpdateRequest;
+import com.example.orderapi.dto.ordergroup.OrderGroupUpdateDeliveryInfoRequest;
+import com.example.orderapi.dto.wrapping.WrappingResponse;
+import com.example.orderapi.entity.deliveryinfo.DeliveryInfo;
 import com.example.orderapi.entity.ordergroup.OrderGroup;
+import com.example.orderapi.entity.wrapping.Wrapping;
 import com.example.orderapi.repository.ordergroup.OrderGroupRepository;
+import com.example.orderapi.service.deliveryinfo.DeliveryInfoService;
+import com.example.orderapi.service.wrapping.WrappingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,15 +19,17 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
-@Service
-@Transactional
 @RequiredArgsConstructor
+@Transactional
+@Service
 public class OrderGroupServiceImpl implements OrderGroupService {
 
     private final OrderGroupRepository orderGroupRepository;
+    private final DeliveryInfoService deliveryInfoService;
+    private final WrappingService wrappingService;
 
     @Override
-    public OrderGroupResponse getById(Long id) {
+    public OrderGroupResponse getOrderGroupById(Long id) {
         Optional<OrderGroup> optionalOrderGroup = orderGroupRepository.findById(id);
 
         if (optionalOrderGroup.isEmpty()) {
@@ -32,20 +40,46 @@ public class OrderGroupServiceImpl implements OrderGroupService {
     }
 
     @Override
-    public Page<OrderGroupResponse> getOrderGroupsByUserId(Long userId, Pageable pageable) {
+    public Page<OrderGroupResponse> getOrderGroupPagesByUserId(Long userId, Pageable pageable) {
         Page<OrderGroup> orderGroups = orderGroupRepository.findAllByUserId(userId, pageable);
         return orderGroups.map(OrderGroupResponse::fromEntity);
     }
 
     @Override
-    public OrderGroupResponse update(Long id, OrderGroupUpdateRequest orderGroupUpdateRequest) {
+    public OrderGroupResponse updateOrderGroup(Long id, OrderGroupUpdateDeliveryInfoRequest OrderGroupUpdateDeliveryInfoRequest) {
         Optional<OrderGroup> optionalOrderGroup = orderGroupRepository.findById(id);
 
         if (optionalOrderGroup.isEmpty()) {
             throw new RuntimeException();
         }
+
         OrderGroup orderGroup = optionalOrderGroup.get();
-        orderGroup.setDeliveryInfoId(orderGroupUpdateRequest.getDeliveryInfoId());
+
+        DeliveryInfoResponse deliveryInfoResponse = deliveryInfoService.getDeliveryInfoById(OrderGroupUpdateDeliveryInfoRequest.getDeliveryInfoId());
+        DeliveryInfo deliveryInfo = new DeliveryInfo();
+
+        deliveryInfo.ofCreate(deliveryInfoResponse.getName(), deliveryInfoResponse.getInvoiceNumber());
+
+        orderGroup.ofUpdateDeliveryInfo(deliveryInfo);
+
+        return OrderGroupResponse.fromEntity(orderGroup);
+    }
+
+    @Override
+    public OrderGroupResponse createOrderGroup(OrderGroupCreateRequest orderGroupCreateRequest) {
+        OrderGroup orderGroup = new OrderGroup();
+
+        WrappingResponse wrappingResponse = wrappingService.getWrappingById(orderGroupCreateRequest.getWrappingId());
+        Wrapping wrapping = new Wrapping();
+        wrapping.ofCreate(wrappingResponse.getName(), wrappingResponse.getPrice());
+
+        orderGroup.ofCreate(
+                orderGroupCreateRequest.getUserId(),
+                orderGroupCreateRequest.getOrderedName(),
+                orderGroupCreateRequest.getRecipientName(),
+                orderGroupCreateRequest.getRecipientPhone(),
+                orderGroupCreateRequest.getDeliveryPrice(),
+                wrapping);
 
         orderGroupRepository.save(orderGroup);
 
@@ -53,23 +87,10 @@ public class OrderGroupServiceImpl implements OrderGroupService {
     }
 
     @Override
-    public OrderGroupResponse create(OrderGroupCreateRequest request) {
-        OrderGroup OrderGroup = new OrderGroup();
-
-        OrderGroup.setUserId(request.getUserId());
-        OrderGroup.setWrappingId(request.getWrappingId());
-        OrderGroup.setOrderedAt(request.getOrderedAt());
-        OrderGroup.setRecipientName(request.getRecipientName());
-        OrderGroup.setRecipientPhone(request.getRecipientPhone());
-        OrderGroup.setDeliveryPrice(request.getDeliveryPrice());
-
-        orderGroupRepository.save(OrderGroup);
-
-        return OrderGroupResponse.fromEntity(OrderGroup);
-    }
-
-    @Override
-    public void delete(Long id) {
+    public void deleteOrderGroup(Long id) {
+        if (!orderGroupRepository.existsById(id)) {
+            throw new RuntimeException();
+        }
         orderGroupRepository.deleteById(id);
     }
